[cpp] Adjust 10x indices to int64_t

+++ r/src/bpcells-cpp/matrixIterators/ImportMatrix10xHDF5.cpp
@@ -100,7 +100,7 @@ StoredMatrixWriter<T> create10xFeatureMatrix(
     wb.createStringWriter("features/_all_tag_keys")->write(VecStringReader(tag_keys));
 
     return StoredMatrixWriter(
-        wb.create<int32_t>("indices").convert<uint32_t>(),
+        wb.create<int64_t>("indices").convert<uint32_t>(),
         wb.create<T>("data"),
         wb.create<int64_t>("indptr").convert<uint64_t>(),
         wb.create<int32_t>("shape").convert<uint32_t>(),
